+++ wn/_add.py
@@ -322,8 +322,9 @@ def _insert_lexicon(
 ) -> tuple[int, int]:
     progress.set(status='Lexicon Info')
     cur.execute(
-        'INSERT INTO lexicons VALUES (null,?,?,?,?,?,?,?,?,?,?,?)',
-        (lexicon['id'],
+        'INSERT INTO lexicons VALUES (null,?,?,?,?,?,?,?,?,?,?,?,?)',
+        (f"{lexicon['id']}:{lexicon['version']}",
+         lexicon['id'],
          lexicon['label'],
          lexicon['language'],
          lexicon['email'],
@@ -420,7 +421,7 @@ def _insert_synsets(
     # synsets
     ss_query = f'''
         INSERT INTO synsets
-        VALUES (null,?,?,(SELECT rowid FROM ilis WHERE id=?),?,?,({LEXFILE_QUERY}),?)
+        VALUES (null,?,?,(SELECT rowid FROM ilis WHERE id=?),?,({LEXFILE_QUERY}),?)
     '''
     # presupposed ILIs
     pre_ili_query = f'''
@@ -457,7 +458,6 @@ def _insert_synsets(
              lexid,
              ss['ili'] if ss['ili'] and ss['ili'] != 'in' else None,
              ss['partOfSpeech'],
-             ss.get('lexicalized', True),
              ss.get('lexfile'),
              ss['meta'])
             for ss in batch
@@ -476,6 +476,18 @@ def _insert_synsets(
         cur.executemany(pro_ili_query, pro_ili_data)
 
         progress.update(len(batch))
+
+    # only store when lexicalized=False
+    unlexicalized_data = [
+        (synset['id'], lexid)
+        for synset in _local_synsets(synsets)
+        if not synset.get('lexicalized', True)
+    ]
+    query = f'''
+        INSERT INTO unlexicalized_synsets (synset_rowid) {SYNSET_QUERY}
+    '''
+    cur.executemany(query, unlexicalized_data)
+
 
 
 def _insert_synset_definitions(
@@ -677,7 +689,6 @@ def _insert_senses(
                 ?,
                 ({SYNSET_QUERY}),
                 ?,
-                ?,
                 ?)
     '''
     for batch in _batch(entries):
@@ -688,13 +699,24 @@ def _insert_senses(
              i,
              sense['synset'], lexidmap.get(sense['synset'], lexid),
              ssrank.get(sense['id'], DEFAULT_MEMBER_RANK),
-             sense.get('lexicalized', True),
              sense['meta'])
             for entry in batch
             for i, sense in enumerate(_local_senses(_senses(entry)))
         ]
         cur.executemany(query, data)
         progress.update(len(data))
+
+    # only store when lexicalized=False
+    unlexicalized_data = [
+        (sense['id'], lexid)
+        for entry in entries
+        for sense in _local_senses(_senses(entry))
+        if not sense.get('lexicalized', True)
+    ]
+    query = f'''
+        INSERT INTO unlexicalized_senses (sense_rowid) {SENSE_QUERY}
+    '''
+    cur.executemany(query, unlexicalized_data)
 
 
 def _insert_adjpositions(
@@ -952,7 +974,7 @@ def remove(
                 for ext_spec in reversed(extensions):
                     progress.set(status=f'{ext_spec} (extension)')
                     conn.execute(
-                        'DELETE FROM lexicons WHERE id || ":" || version = ?',
+                        'DELETE FROM lexicons WHERE specifier = ?',
                         (ext_spec,),
                     )
                     progress.flash(f'Removed {ext_spec}\n')
@@ -960,7 +982,7 @@ def remove(
                 extra = f' (and {len(extensions)} extension(s))' if extensions else ''
                 progress.set(status=f'{lexspec}', count=0)
                 conn.execute(
-                    'DELETE FROM lexicons WHERE id || ":" || version = ?',
+                    'DELETE FROM lexicons WHERE specifier = ?',
                     (lexspec,),
                 )
                 progress.flash(f'Removed {lexspec}{extra}\n')