+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased][unreleased]
 
+### Schema
+
+* Add `specifier` column to `lexicon` table ([#234])
+* Remove `lexicalized` column from `synsets` and `senses` ([#248])
+* Add `unlexicalized_synsets` and `unlexicalized_senses` tables ([#248])
+
+### Changed
+
+* Default form normalizer uses casefold instead of lower ([#233])
+
 
 ## [v0.12.0]
 
@@ -766,8 +776,11 @@ abandoned, but this is an entirely new codebase.
 [#221]: https://github.com/goodmami/wn/issues/221
 [#226]: https://github.com/goodmami/wn/issues/226
 [#228]: https://github.com/goodmami/wn/issues/228
+[#233]: https://github.com/goodmami/wn/issues/233
+[#234]: https://github.com/goodmami/wn/issues/234
 [#235]: https://github.com/goodmami/wn/issues/235
 [#238]: https://github.com/goodmami/wn/issues/238
 [#241]: https://github.com/goodmami/wn/issues/241
 [#246]: https://github.com/goodmami/wn/issues/246
+[#248]: https://github.com/goodmami/wn/issues/248
 [#250]: https://github.com/goodmami/wn/issues/250